+++ lib/printer/CMakeLists.txt
@@ -1,0 +1,11 @@
+add_library(KOREPrinter
+  addBrackets.cpp
+  printer.cpp
+)
+
+install(
+  TARGETS KOREPrinter
+  ARCHIVE DESTINATION lib/kllvm
+)
+
+target_link_libraries(KOREPrinter PUBLIC Parser AST)
